% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_inverse.R
\name{coef.lm_pseudo}
\alias{coef.lm_pseudo}
\title{Extract Coefficients from lm_pseudo Objects}
\usage{
\method{coef}{lm_pseudo}(object)
}
\arguments{
\item{object}{An object of class \code{lm_pseudo}.}
}
\value{
A named numeric vector of estimated coefficients.
}
\description{
Retrieves the estimated coefficients from an \code{lm_pseudo} object.
}
