% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex.R
\name{simplex_project}
\alias{simplex_project}
\title{Project a Vector onto the Simplex}
\usage{
simplex_project(x, z = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{z}{Simplex radius. Defaults to 1.}
}
\value{
Numeric vector on the simplex.
}
\description{
Project a Vector onto the Simplex
}
