% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{simulate_ar_l_with_initial}
\alias{simulate_ar_l_with_initial}
\title{Simulate Time Series from AR(l) Model with Specific Initial Values}
\usage{
simulate_ar_l_with_initial(phi, y0, n, sd_eps)
}
\arguments{
\item{phi}{A numeric vector of length \code{l} containing the autoregressive coefficients
for each lag. The AR(1) coefficient corresponds to \code{phi[1]}, AR(2) to \code{phi[2]},
and so on up to AR(\code{l}).}

\item{y0}{A numeric vector of length \code{l} specifying the initial values for each lag
of the time series. The value \code{y0[i]} represents the initial value for lag \code{i}.}

\item{n}{An integer specifying the number of observations to simulate.}

\item{sd_eps}{A numeric value representing the standard deviation of the innovations.
The innovations are assumed to be independent and identically distributed (iid) Gaussian white noise
with mean 0 and standard deviation \code{sd_eps}.}
}
\value{
A numeric vector containing the simulated time series.
}
\description{
This function simulates a time series from an autoregressive model of order l (AR(l))
with specified initial values for each lag.
}
\examples{
# Simulate an AR(2) series with specific initial values
phi <- c(0.3, 0.2) # AR(2) coefficients
y0 <- c(10, 15) # Initial values for each lag
n <- 100 # Number of observations
sd_eps <- 1 # Standard deviation of innovations
simulated_series <- simulate_ar_l_with_initial(phi, y0, n, sd_eps)

# Plot the simulated series
plot(simulated_series,
  type = "l", main = "Simulated AR(2) Series with Specific Initial Values",
  xlab = "Time", ylab = "Value"
)

}
\seealso{
\code{\link[=generate_synth_data_ar]{generate_synth_data_ar()}}
}
