% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synth_data}
\alias{synth_data}
\title{Synthetic Data Generated with Interactive Fixed Effect Model}
\format{
A tibble with 100 rows and 19 columns
where \eqn{t0 = 99}, \eqn{n = 16}, and the true ATT is \eqn{0}:
\itemize{
\item \code{year}: A numeric vector of year
\item \code{Dt}: A numeric vector of treatment indicator
\item \code{Y}: A numeric vector of outcome for West Germany
\item \code{X1} to \code{X16}: Numeric vectors of outcome for control units
}
}
\usage{
synth_data
}
\description{
This data set is generated with the function \code{generate_synth_data_ife}.
Specifically, using the panel data from Abadie, Diamond, and Hainmueller (2015),
we fit the interactive fixed effect model and generate synthetic data
with the estimated parameters.
}
\keyword{datasets}
