% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synth_post}
\alias{synth_post}
\title{Synthetic Postperiod Data}
\format{
A tibble with 1 row and 17 columns, where the columns include the following:
\itemize{
\item \code{Y}: A numeric vector of outcome for treated unit
\item \code{X1} to \code{X16}: Numeric vectors of outcome for control units
}
}
\usage{
synth_post
}
\description{
A wide-format panel data set of synthetic postperiod data.
}
\keyword{datasets}
