% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synth_pre}
\alias{synth_pre}
\title{Synthetic Preperiod Data}
\format{
A tibble with 99 rows and 17 columns, where the columns include the following:
\itemize{
\item \code{Y}: A numeric vector of outcome for treated unit
\item \code{X1} to \code{X16}: Numeric vectors of outcome for control units
}
}
\usage{
synth_pre
}
\description{
A wide-format panel data set of synthetic preperiod data.
}
\keyword{datasets}
