% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taiwan}
\alias{taiwan}
\title{Taiwan’s Expulsion from the IMF Data (Lipscy and Lee 2019)}
\format{
A tibble with 21 rows and 77 columns, where the columns include the following:
\itemize{
\item \code{year}: A numeric vector of year
\item \code{Dt}: A numeric vector of treatment indicator
\item \code{Taiwan}: A numeric vector of outcome for Taiwan
\item The rest of the columns are numeric vectors of outcome for control units
}
}
\source{
\url{https://static.cambridge.org/content/id/urn:cambridge.org:id:article:S0020818318000371/resource/name/S0020818318000371sup001.zip}
}
\usage{
taiwan
}
\description{
A wide-format panel data set of Taiwan's expulsion from the International Monetary Fund (IMF) from Lipscy and Lee (2019).
Note that it does include missing values coded as \code{NA}.
}
\keyword{datasets}
