% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgermany}
\alias{wgermany}
\title{West German Reunification Data (Abadie, Diamond, and Hainmueller 2015)}
\format{
A tibble with 44 rows and 19 columns, where the columns include the following:
\itemize{
\item \code{year}: A numeric vector of year
\item \code{Dt}: A numeric vector of treatment indicator
\item \verb{West Germany}: A numeric vector of outcome for West Germany
\item The rest of the columns are numeric vectors of outcome for control units
}
}
\source{
\url{https://doi.org/10.7910/DVN/24714}
}
\usage{
wgermany
}
\description{
A wide-format panel data set of West German reunification data from Abadie, Diamond, and Hainmueller (2014).
Note that it does not include missing values.
}
\keyword{datasets}
