% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_params.R
\name{estimate_params}
\alias{estimate_params}
\title{Leave-one-out Analysis}
\usage{
estimate_params(
  var_y_name,
  var_x_name,
  data_pre,
  data_post,
  pseudo_inverse = FALSE
)
}
\arguments{
\item{var_y_name}{Name of the treated unit.}

\item{var_x_name}{Names of the control units.}

\item{data_pre}{Pre-period data where the rows are time points and the columns include treated/control units. See \code{synth_pre} for example.}

\item{data_post}{Post-period data. Should have identical columns to data_pre. See \code{synth_post} for example.}

\item{pseudo_inverse}{A logical to indicate whether to use the pseudo-inverse to fit the model.}
}
\value{
A data.frame with the following columns:
\itemize{
\item \code{dropped_unit}: Name of the control unit that is dropped.
\item \code{time_from_treatment}: Time from the treatment.
\item \code{estimate}: Estimated treatment effect.
\item \code{std.error}: Standard error of the treatment effect (based on the vertical regression).
\item \code{statistic}: t-statistic of the treatment effect.
\item \code{p.value}: p-value of the treatment effect.
\item \code{gamma}: Weight of the control unit.
\item \code{delta}: Imbalance of the control unit.
\item \code{bias}: Bias of the treatment effect.
\item \code{tau}: Estimated treatment effect with the entire data.
\item \code{tau_plus_bias}: Sum of the estimated treatment effect and the bias. Must be close to the estimate (= estimate + noise).
\item \code{r2_Y_Z}: Partial R squared of the control unit with the outcome.
\item \code{r2_D_Z}: Partial R squared of the control unit with the treatment indicator.
\item \code{tau_se}: Standard error of the treatment effect with the entire data.
\item \code{t_stat}: t-Statistic of the treatment effect with the entire data.
}
}
\description{
Function to estimate bias parameters with dropping one control unit at a time
}
\examples{
estimate_params(
  var_y_name = "Y",
  var_x_name = paste0("X", 1:16),
  data_pre = synth_pre,
  data_post = synth_post,
  pseudo_inverse = FALSE
)

}
\seealso{
\code{\link[=estimate_params_partial]{estimate_params_partial()}}, \code{\link[=estimate_params_partial_multi]{estimate_params_partial_multi()}}
}
