% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_params.R
\name{estimate_params_partial_multi}
\alias{estimate_params_partial_multi}
\title{Estimate Sensitivity Parameters with Multiple Control Units with Partially Observed Data}
\usage{
estimate_params_partial_multi(
  var_z_name,
  var_x_name,
  var_y_name,
  data_pre,
  data_post,
  pseudo_inverse = FALSE
)
}
\arguments{
\item{var_z_name}{A vector of strings specifying the name of the control unit.}

\item{var_x_name}{A vector of strings specifying the name of the control unit.}

\item{var_y_name}{Name of the treated unit.}

\item{data_pre}{Pre-period data where the rows are time points and the columns include treated/control units. See \code{synth_pre} for example.}

\item{data_post}{Post-period data. Should have identical columns to data_pre. See \code{synth_post} for example.}

\item{pseudo_inverse}{A logical to indicate whether to use the pseudo-inverse to fit the model.}
}
\value{
A list of \code{gamma}, \code{imbalance}, and \code{bias}.
}
\description{
Function to use partially observed data to estimate gamma and imbalance,
in case of multiple control units with missing observations.
See \code{\link[=estimate_params_partial]{estimate_params_partial()}} for a single control unit.
}
\examples{
# Suppose X1 and X2 are partially observed
synth_pre_partial <- synth_pre |>
  dplyr::mutate(
    X1 = ifelse(dplyr::row_number() < dplyr::n() / 2, NA, X1),
    X2 = ifelse(dplyr::row_number() < dplyr::n() / 2, NA, X2)
  ) |>
  dplyr::filter(!is.na(X1) | is.na(X1))
estimate_params_partial_multi(
  var_z_name = c("X1", "X2"),
  var_x_name = paste0("X", 3:16),
  var_y_name = "Y",
  data_pre = synth_pre_partial,
  data_post = synth_post,
  pseudo_inverse = FALSE
)

}
\seealso{
\code{\link[=estimate_params]{estimate_params()}}, \code{\link[=estimate_params_partial]{estimate_params_partial()}}
}
