% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_synth_data_ar}
\alias{generate_synth_data_ar}
\title{Synthetic Data with Autoregressive Model AR(l)}
\usage{
generate_synth_data_ar(
  t0 = 99,
  n = 20,
  tau = 10,
  ar = c(0.7, 0.2, -0.1, -0.3),
  init = NULL
)
}
\arguments{
\item{t0}{Numeric, the number of pre-treatment time periods (observations).}

\item{n}{Numeric, the number of donor units (variables).}

\item{tau}{Numeric, the value of ATT (Average Treatment Effect on the Treated).}

\item{ar}{Numeric vector of length l, coefficients for AR model.}

\item{init}{(n+1) by l matrix, initial values for AR model (optional).}
}
\value{
A list containing the following elements:
\itemize{
\item df_pre_full: A data frame of complete pre-treatment data (t0 x n+1). The last column is the observation for treated unit (Y).
\item df_prepost_full: A data frame of complete pre- and post-treatment data ((t0+1) x n+1). The second to last column is the observation for treated unit (Y), and the last column is the treatment indicator (D). The last row is the post-treatment observation.
\item tau: A numeric value of ATT
\item .call: The matched call.
}
}
\description{
Synthetic Data with Autoregressive Model AR(l)
}
\examples{
set.seed(123)
synth <- generate_synth_data_ar(t0 = 99, n = 20, tau = 10, ar = c(0.7, 0.2, -0.1, -0.3))

}
\seealso{
\code{\link[=simulate_ar_l_with_initial]{simulate_ar_l_with_initial()}}
}
