% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{generate_synth_data_vertreg}
\alias{generate_synth_data_vertreg}
\title{Synthetic Data with Vanilla Vertical Regression}
\usage{
generate_synth_data_vertreg(
  t0 = 99,
  n = 20,
  tau = 10,
  rho = 0.04,
  mu = NULL,
  beta = NULL
)
}
\arguments{
\item{t0}{Numeric, the number of pre-treatment time periods (observations).}

\item{n}{Numeric, the number of donor units (variables).}

\item{tau}{Numeric, the value of ATT (Average Treatment Effect on the Treated).}

\item{rho}{Numeric, a parameter for non-overlap.}

\item{mu}{Numeric, another parameter for non-overlap.}

\item{beta}{Numeric vector, weights.}
}
\value{
A list containing the following elements:
- df_pre_full: A data frame of complete pre-treatment data (t0 x n+1). The last column is the observation for treated unit (Y).
- df_prepost_full: A data frame of complete pre- and post-treatment data ((t0+1) x n+1). The second to last column is the observation for treated unit (Y), and the last column is the treatment indicator (D). The last row is the post-treatment observation.
- beta: A numeric vector of weights.
- tau: A numeric value of ATT
- .call: The matched call.
}
\description{
Generates synthetic data based on a vanilla vertical regression model given a set of parameters.
This method is designed to analyze data with a specified number of pre-treatment time periods and donor units,
incorporating an ATT (Average Treatment Effect on the Treated) value and parameters for non-overlap.
}
\examples{
set.seed(123)
synth <- generate_synth_data_vertreg(t0 = 99, n = 20, tau = 10, rho = 0.04, mu = NULL, beta = NULL)

}
