% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_inverse.R
\name{lm_pseudo}
\alias{lm_pseudo}
\title{Fitting Linear Models Using Pseudo-inverse}
\usage{
lm_pseudo(formula, data)
}
\arguments{
\item{formula}{A string to specify the linear model.}

\item{data}{A data.frame containing the data.}
}
\value{
A list containing the estimated coefficients.
}
\description{
Function to fit linear models using the pseudo-inverse.
}
