% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Sensitivity Contour Plot}
\usage{
plot_sensitivity(
  tau_complete,
  var_x_gamma,
  var_x_delta,
  var_x_name,
  title = NULL,
  gamma_seq = NULL,
  delta_seq = NULL,
  text_size = 3,
  nudge_x = 0.02,
  nudge_y = 0,
  repel = FALSE,
  tau_round = 1,
  nudge_x2 = NULL,
  nudge_y2 = NULL
)
}
\arguments{
\item{tau_complete}{Estimated treatment effect with the entire data.}

\item{var_x_gamma}{Weights of the control units.}

\item{var_x_delta}{Imbalance of the control units.}

\item{var_x_name}{Names of the control units.}

\item{title}{Title of the plot}

\item{gamma_seq}{Sequence of gamma values to plot.}

\item{delta_seq}{Sequence of delta values to plot.}

\item{text_size}{Size of the text}

\item{nudge_x}{Nudge value for the x-axis text}

\item{nudge_y}{Nudge value for the y-axis text}

\item{repel}{Whether to use \code{ggrepel::geom_text_repel()} for text labels.}

\item{tau_round}{Number of decimal places to round the treatment effect.}

\item{nudge_x2}{Nudge value for the x-axis text of the treatment effect.}

\item{nudge_y2}{Nudge value for the y-axis text of the treatment effect.}
}
\value{
ggplot2 object
}
\description{
Generate a sensitivity contour plot with the estimated treatment effect, weights, and imbalance.
}
\examples{
# leave-one-out sensitivity analysis
res <- estimate_params(var_y_name = "Y",
                       var_x_name = paste0("X", 1:16),
                       data_pre = synth_pre,
                       data_post = synth_post)
tau_complete <- res |> dplyr::pull(tau) |> unique()
var_x_gamma <- res |> dplyr::pull(gamma)
var_x_delta <- res |> dplyr::pull(delta)
var_x_tau <- res |> dplyr::pull(estimate)
var_x_name <- paste0("Drop ", paste0("X", 1:16), " (", round(var_x_tau, 1), ")")
plot_sensitivity(tau_complete = tau_complete,
                 var_x_gamma = var_x_gamma, var_x_delta = var_x_delta, var_x_name = var_x_name,
                 title = "Sensitivity of the SCM Estimate", text_size = 4)


}
\seealso{
\code{\link[=plot_sensitivity_r2]{plot_sensitivity_r2()}}
}
