% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sensitivity_r2}
\alias{plot_sensitivity_r2}
\title{Sensitivity Contour Plot with Partial R Squared}
\usage{
plot_sensitivity_r2(
  tau_complete,
  df,
  tau_se,
  sign_bias,
  var_x_r2_Y_Z,
  var_x_r2_D_Z,
  var_x_name,
  title = NULL,
  r2_Y_Z_seq = seq(0, 1, length.out = 100),
  r2_D_Z_seq = seq(0, 1, length.out = 100),
  text_size = 3,
  plot_t_stat = FALSE,
  t_stat = NULL,
  critical_value = NULL
)
}
\arguments{
\item{tau_complete}{Estimated treatment effect with the entire data.}

\item{df}{Degree of freedom of the vertical regression with the entire data.}

\item{tau_se}{Standard error of the treatment effect with the entire data.}

\item{sign_bias}{Sign of the bias.}

\item{var_x_r2_Y_Z}{Partial R squared of the control units with the outcome.}

\item{var_x_r2_D_Z}{Partial R squared of the control units with the treatment indicator.}

\item{var_x_name}{Names of the control units.}

\item{title}{Title of the plot}

\item{r2_Y_Z_seq}{Sequence of x values to plot.}

\item{r2_D_Z_seq}{Sequence of y values to plot.}

\item{text_size}{Size of the text}

\item{plot_t_stat}{Whether to plot t-statistic instead of bias. If true, \code{t_stat} argument should be provided.}

\item{t_stat}{t-statistic of the treatment effect with the entire data.}

\item{critical_value}{Critical value for the t-statistic contour line.}
}
\value{
ggplot2 object
}
\description{
Generate a sensitivity contour plot of partial R squared with the estimated treatment effect, weights, and imbalance.
}
\examples{
# leave-one-out sensitivity analysis
fit <- vertreg_stacked(formula = paste("Y ~ -1 + ",
                                       paste(paste0("X", 1:16),
                                             collapse = " + ")),
                       data_pre = synth_pre,
                       data_post = synth_post,
                       pseudo_inverse = FALSE)
df <- fit$df
tau_se <- fit$std.error
res <- estimate_params(var_y_name = "Y",
                       var_x_name = paste0("X", 1:16),
                       data_pre = synth_pre,
                       data_post = synth_post)
# plot only top 5 units with positive bias
res <- res[order(-res$bias), ][1:5, ]
tau_complete <- res |> dplyr::pull(tau) |> unique()
var_x_r2_Y_Z <- res |> dplyr::pull(r2_Y_Z)
var_x_r2_D_Z <- res |> dplyr::pull(r2_D_Z)
var_x_name <- paste0("Drop ", paste0("X", 1:5))
plot_sensitivity_r2(tau_complete = tau_complete, df = df, tau_se = tau_se,
                    sign_bias = 1,
                    var_x_r2_Y_Z = var_x_r2_Y_Z, var_x_r2_D_Z = var_x_r2_D_Z,
                    var_x_name = var_x_name,
                    title = "Sensitivity of the SCM Estimate (Positive Bias)",
                    r2_Y_Z_seq = seq(0, 0.1, length.out = 100),
                    r2_D_Z_seq = seq(0, 0.1, length.out = 100),
                    text_size = 5)

}
\seealso{
\code{\link[=plot_sensitivity]{plot_sensitivity()}}
}
