% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertreg.R
\name{vertreg}
\alias{vertreg}
\title{Function to run a vertical regression.}
\usage{
vertreg(formula, data_pre, data_post)
}
\arguments{
\item{formula}{A string to specify the vertical regression model (e.g., "Y ~ X1 + X2" where "Y" is treated unit, "X1" and "X2" are control units).}

\item{data_pre}{Pre-period data where the rows are time points and the columns include treated/control units.}

\item{data_post}{Post-period data. Should have identical columns to data_pre.}
}
\value{
A vector of treatment effects.
}
\description{
Function to run a vertical regression.
}
\examples{
df_pre_full <- synth_data |>
  dplyr::filter(Dt == 0) |>
  dplyr::select(-year, -Dt)
df_post_full <- synth_data |>
  dplyr::filter(Dt == 1) |>
  dplyr::select(-year, -Dt)
formula_full <- as.formula(
  paste("Y ~ -1 + ", paste(paste0("X", 1:16), collapse = " + "))
)
vertreg(
  formula = formula_full,
  data_pre = df_pre_full,
  data_post = df_post_full
)

}
