% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertreg.R
\name{vertreg_stacked}
\alias{vertreg_stacked}
\title{Vertical Regression (With Treatment Indicator)}
\usage{
vertreg_stacked(formula, data_pre, data_post, pseudo_inverse = FALSE)
}
\arguments{
\item{formula}{A string to specify the vertical regression model (e.g., "Y ~ X1 + X2" where "Y" is treated unit, "X1" and "X2" are control units).}

\item{data_pre}{Pre-period data where the rows are time points and the columns include treated/control units. See \code{synth_pre} for example.}

\item{data_post}{Post-period data. Should have identical columns to data_pre. See \code{synth_post} for example.}

\item{pseudo_inverse}{A logical to indicate whether to use the pseudo-inverse to fit the model.}
}
\value{
A data.frame with the treatment effect and its standard error based on the vertical regression:
\itemize{
\item \code{time_from_treatment}: the time from the treatment
\item \code{estimate}: the estimated treatment effect
\item \code{std.error}: the standard error of the estimated treatment effect
\item \code{statistic}: the t-statistic of the estimated treatment effect
\item \code{p.value}: the p-value of the estimated treatment effect
\item \code{df}: the degree of freedom of the regression
}

@seealso \code{\link[=vertreg]{vertreg()}}
}
\description{
Function to run a vertical regression in stacked form with a treatment indicator.
}
\examples{
vertreg_stacked(
  formula = paste("Y ~ -1 + ", paste(paste0("X", 1:16), collapse = " + ")),
  data_pre = synth_pre,
  data_post = synth_post
)

}
