% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertreg.R
\name{vertreg_stacked}
\alias{vertreg_stacked}
\title{Vertical Regression (With Treatment Indicator)}
\usage{
vertreg_stacked(formula, data_pre, data_post)
}
\arguments{
\item{formula}{A string to specify the vertical regression model (e.g., "Y ~ X1 + X2" where "Y" is treated unit, "X1" and "X2" are control units).}

\item{data_pre}{Pre-period data where the rows are time points and the columns include treated/control units.}

\item{data_post}{Post-period data. Should have identical columns to data_pre.}
}
\value{
A data.frame with the treatment effect and its standard error based on the vertical regression:
\itemize{
\item \code{time_from_treatment}: the time from the treatment
\item \code{estimate}: the estimated treatment effect
\item \code{std.error}: the standard error of the estimated treatment effect
\item \code{statistic}: the t-statistic of the estimated treatment effect
\item \code{p.value}: the p-value of the estimated treatment effect
\item \code{df}: the degree of freedom of the regression
}
}
\description{
Function to run a vertical regression in stacked form with a treatment indicator.
}
\examples{
df_pre_full <- synth_data |>
  dplyr::filter(Dt == 0) |>
  dplyr::select(-year, -Dt)
df_post_full <- synth_data |>
  dplyr::filter(Dt == 1) |>
  dplyr::select(-year, -Dt)
formula_full <- as.formula(
  paste("Y ~ -1 + ", paste(paste0("X", 1:16), collapse = " + "))
)
vertreg_stacked(
  formula = formula_full,
  data_pre = df_pre_full,
  data_post = df_post_full
)

}
